+++ mediaframe/src/serde_impls/mod.rs
@@ -252,6 +252,10 @@ serde_via_str!(crate::codec::AudioCodec);
 #[cfg(any(feature = "std", feature = "alloc"))]
 serde_via_str!(crate::codec::SubtitleCodec);
 #[cfg(any(feature = "std", feature = "alloc"))]
+serde_via_str!(crate::codec::DataCodec);
+#[cfg(any(feature = "std", feature = "alloc"))]
+serde_via_str!(crate::codec::AttachmentCodec);
+#[cfg(any(feature = "std", feature = "alloc"))]
 serde_via_str!(crate::container::Format);
 #[cfg(any(feature = "std", feature = "alloc"))]
 serde_via_str!(crate::image::Format);